% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cp_path}
\alias{cp_path}
\title{Create path relative to root of project}
\usage{
cp_path(path)
}
\arguments{
\item{path}{Path to be appended to project root}
}
\description{
Create path relative to root of project
}
