% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pastellize}
\alias{pastellize}
\title{Create pastel color palatte}
\usage{
pastellize(x, p)
}
\arguments{
\item{x}{hsv value}

\item{p}{Level of pastellisation. Nnumber in \link{0,1}}
}
\description{
Create pastel color palatte
}
