% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_figs}
\alias{save_figs}
\title{Save Figures}
\usage{
save_figs(
  name,
  fig,
  width = 6,
  height = 6,
  plot_dir = file.path(here::here(), "analysis/plots"),
  pdf_plot = TRUE,
  font_family = "Helvetica",
  dpi = 600
)
}
\arguments{
\item{name}{Name of figure}

\item{fig}{ggplot or similar figure object}

\item{width}{Width of plot in inches. Default = 6}

\item{height}{Height of plot in inches. Default = 6}

\item{plot_dir}{Plotting directory. Defaults to "analysis/plots"}

\item{pdf_plot}{Logical for plotting pdf too. Default = TRUE}

\item{font_family}{If specified, sets all font family. Default = NULL}

\item{dpi}{DPI to use for raster graphics. Default = 600}
}
\description{
Save Figures
}
