% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_mapping.R
\name{who_compliant_continuous_plot}
\alias{who_compliant_continuous_plot}
\title{Generate WHO-Compliant Continuous Plot}
\usage{
who_compliant_continuous_plot(
  data,
  var,
  title,
  who_shps,
  region = "africa",
  limits = FALSE,
  print_bool = FALSE,
  lakes_bool = TRUE,
  disputed_areas_bool = TRUE,
  disputed_borders_bool = TRUE,
  scale_bar_bool = FALSE,
  risk = "innate",
  prev_plot = FALSE,
  prev_var = "Micro.2.10_mean",
  scale_fill_func = function(...) {
     scale_fill_viridis_c(option = "C", direction =
    -1, end = 0.7, ...)
 },
  cut_off = NULL
)
}
\arguments{
\item{data}{A \code{data.frame} containing the variable of interest and corresponding administrative IDs (\code{id_1}) to merge with WHO shapefiles.}

\item{var}{A string specifying the column name in \code{data} to be plotted as the fill aesthetic.}

\item{title}{A string specifying the title for the plot legend.}

\item{who_shps}{A list of WHO shapefiles containing spatial data for administrative boundaries, lakes, and disputed regions.}

\item{region}{A string specifying the region of interest. Options are \code{"global"}, \code{"africa"}, \code{"asia"}, or \code{"latam"}. Defaults to \code{"africa"}.}

\item{limits}{Logical. If \code{TRUE}, limits the plotted data to the specified region. Defaults to \code{FALSE}.}

\item{print_bool}{Logical. If \code{TRUE}, prints the generated map. Defaults to \code{FALSE}.}

\item{lakes_bool}{Logical. If \code{TRUE}, includes lakes in the map. Defaults to \code{TRUE}.}

\item{disputed_areas_bool}{Logical. If \code{TRUE}, highlights disputed areas on the map. Defaults to \code{TRUE}.}

\item{disputed_borders_bool}{Logical. If \code{TRUE}, outlines disputed borders on the map. Defaults to \code{TRUE}.}

\item{scale_bar_bool}{Logical. If \code{TRUE}, adds a scale bar to the map. Defaults to \code{FALSE}.}

\item{prev_plot}{Logical. If \code{TRUE}, layers the new map on top of a previous plot. Defaults to \code{FALSE}.}

\item{prev_var}{Character. Variable name in data for malaria prevalence argument. Defaults to \code{Micro.2.10_mean}.}

\item{scale_fill_func}{Function. Function to call for continuous scale function.
Defaults to \code{function(...){scale_fill_viridis_c(option = "C", direction = 1, end = 0.7, ...)}}.}

\item{cut_off}{List. If not null, this list used to add a layer based on cut offs for the main aesthetics. List
should have named elements \code{c("breaks", "limits", "color", "cutoff", "name", "label")}. Defaults to \code{NULL}}
}
\value{
A \code{ggplot2} object representing the map, which can be printed or further modified. If \code{print_bool} is \code{TRUE}, the map is also printed.
}
\description{
This function generates a geographic map plot displaying WHO-compliant continuous cutoffs for a specified variable across administrative regions. The map can be customised to show specific geographic regions, highlight disputed areas and borders, and include lakes or scale bars.
}
\details{
\itemize{
\item Administrative level 1 regions are merged with the provided data for plotting.
\item Administrative level 0 (country-level) boundaries are overlaid for context.
\item Options to include lakes, disputed areas, and disputed borders enhance the map’s contextual accuracy.
\item The function supports highlighting regions where the variable exceeds specific thresholds.
}
}
\examples{
\dontrun{
who_compliant_continuous_cutoff_plot(
  data = my_data,
  var = "disease_prevalence",
  title = "Prevalence (\%)",
  who_shps = my_who_shapefiles,
  region = "africa",
  lakes_bool = TRUE,
  disputed_areas_bool = TRUE,
  disputed_borders_bool = TRUE
)
}

}
