% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_mapping.R
\name{who_compliant_discrete_plot}
\alias{who_compliant_discrete_plot}
\title{Generate WHO-Compliant Discrete Plot}
\usage{
who_compliant_discrete_plot(
  data,
  var,
  title,
  who_shps,
  region = "africa",
  limits = FALSE,
  print_bool = FALSE,
  lakes_bool = TRUE,
  disputed_areas_bool = TRUE,
  disputed_borders_bool = TRUE,
  scale_bar_bool = FALSE,
  risk = "innate",
  prev_plot = FALSE,
  prev_var = "Micro.2.10_mean",
  scale_fill_func = function(...) {
     ggplot2::scale_fill_manual(values =
    rev(c("#6eb3deff", "#84e1c2ff", "#f8b675ff", "#fa8284ff")), labels = rev(c("No Data",
    "Marginal", "Slight", "Moderate", "High")), ...)
 }
)
}
\arguments{
\item{data}{A \code{data.frame} containing the variable of interest and corresponding administrative IDs (\code{id_1}) for merging with WHO shapefiles.}

\item{var}{A string specifying the name of the column in \code{data} to be used for filling regions on the map.}

\item{title}{A string specifying the title for the legend of the plot.}

\item{who_shps}{A list of WHO shapefiles that contain spatial data for administrative boundaries, lakes, and disputed regions.}

\item{region}{A string specifying the geographic region of interest. Options include \code{"global"}, \code{"africa"}, \code{"asia"}, or \code{"latam"}. Defaults to \code{"africa"}.}

\item{limits}{Logical. If \code{TRUE}, restricts the map's display to the selected region. Defaults to \code{FALSE}.}

\item{print_bool}{Logical. If \code{TRUE}, prints the generated map directly. Defaults to \code{FALSE}.}

\item{lakes_bool}{Logical. If \code{TRUE}, lakes will be displayed on the map. Defaults to \code{TRUE}.}

\item{disputed_areas_bool}{Logical. If \code{TRUE}, disputed areas are highlighted on the map. Defaults to \code{TRUE}.}

\item{disputed_borders_bool}{Logical. If \code{TRUE}, disputed borders are outlined on the map. Defaults to \code{TRUE}.}

\item{scale_bar_bool}{Logical. If \code{TRUE}, adds a scale bar to the map. Defaults to \code{FALSE}.}

\item{prev_plot}{Logical. If \code{TRUE}, overlays the new map on a previous plot. Defaults to \code{FALSE}.}

\item{prev_var}{Character. Variable name in data for malaria prevalence argument. Defaults to \code{Micro.2.10_mean}.}

\item{scale_fill_func}{Function. Function to call for discrete scale function.
Defaults to \code{function(...){scale_fill_viridis_d(option = "C", direction = 1, end = 0.7, ...)}}.}
}
\value{
A \code{ggplot2} object containing the generated map. If \code{print_bool} is \code{TRUE}, the map is also displayed.
}
\description{
This function creates a geographic map visualisation of a discrete variable across WHO-defined administrative regions. The map can be customised to display specific regions, highlight disputed areas and borders, and optionally include lakes or a scale bar.
}
\details{
\itemize{
\item Merges administrative level 1 spatial data with user-provided data for mapping.
\item Outlines country-level administrative boundaries for visual clarity.
\item Optional visual elements include lakes, disputed areas, disputed borders, and scale bars.
\item Uses a discrete colour scale to represent variable values, with a legend for interpretation.
}
}
\examples{
\dontrun{
who_compliant_discrete_plot(
  data = example_data,
  var = "infection_rate",
  title = "Infection Rate (\%)",
  who_shps = who_shapefiles,
  region = "asia",
  lakes_bool = TRUE,
  disputed_areas_bool = TRUE,
  disputed_borders_bool = TRUE
)
}

}
