% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_stargazer}
\alias{write_stargazer}
\title{Write stargazer model output to file}
\usage{
write_stargazer(x, path, cls = NULL, spl = "\\\\s{2,}")
}
\arguments{
\item{x}{Stargazer model output as text}

\item{path}{Path for stargazer to be written to}

\item{cls}{Number of columns in final stargazer table. Default = NULL,
which works it out based on maximum split size}

\item{spl}{regex string for splitting columns.}
}
\description{
Write stargazer model output to file
}
