% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{R6_hrp2_map}
\alias{R6_hrp2_map}
\title{R6 Class for hrp2 results maps}
\description{
A results map
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-hrp2_map-new}{\code{R6_hrp2_map$new()}}
\item \href{#method-hrp2_map-plot}{\code{R6_hrp2_map$plot()}}
\item \href{#method-hrp2_map-set_map_data}{\code{R6_hrp2_map$set_map_data()}}
\item \href{#method-hrp2_map-set_scenarios}{\code{R6_hrp2_map$set_scenarios()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hrp2_map-new"></a>}}
\if{latex}{\out{\hypertarget{method-hrp2_map-new}{}}}
\subsection{Method \code{new()}}{
Create a new hrp2 map result
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_hrp2_map$new(map, map_data = NULL, scenarios = NULL, map_0 = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{map}}{Map sf object}

\item{\code{map_data}}{Map result data for each scenario}

\item{\code{scenarios}}{Scenarios data frame}

\item{\code{map_0}}{Admin 0 map sf object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{hrp2_map} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hrp2_map-plot"></a>}}
\if{latex}{\out{\hypertarget{method-hrp2_map-plot}{}}}
\subsection{Method \code{plot()}}{
Plot hrp2 risk maps

Plot a map for a given scenario and at a given regional scale.

All scenario parameters must be one of "central", "worst", "best"
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_hrp2_map$plot(
  region = "africa",
  Micro.2.10 = "central",
  ft = "central",
  microscopy.use = "central",
  rdt.nonadherence = "central",
  fitness = "central",
  rdt.det = "central",
  print = FALSE,
  risk = "innate"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{region}}{One of "global", "africa", "asia", "latam"}

\item{\code{Micro.2.10}}{Microscopy prevalence scenario}

\item{\code{ft}}{Treatment seeking rate scenario}

\item{\code{microscopy.use}}{Proportion of total testing by microscopy scenario}

\item{\code{rdt.nonadherence}}{Probability of treating neg. indiviudal scenario}

\item{\code{fitness}}{Relative fitness of hrp2 deleted parasite scenario}

\item{\code{rdt.det}}{Chance of hrp2 deleted parasite yielding pos test scenario}

\item{\code{print}}{Boolean for whether to print the plot as well as return it}

\item{\code{risk}}{What type of risk score to plot, either "innate" (default) or "prospective"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot map object silently
Set map data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hrp2_map-set_map_data"></a>}}
\if{latex}{\out{\hypertarget{method-hrp2_map-set_map_data}{}}}
\subsection{Method \code{set_map_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_hrp2_map$set_map_data(map_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{map_data}}{Map data of selection coefficients for each scenario
Set map data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hrp2_map-set_scenarios"></a>}}
\if{latex}{\out{\hypertarget{method-hrp2_map-set_scenarios}{}}}
\subsection{Method \code{set_scenarios()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_hrp2_map$set_scenarios(scenarios)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scenarios}}{Scenario data frame that matches map_data}
}
\if{html}{\out{</div>}}
}
}
}
