% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_model.R
\name{R6_hrp2_mod}
\alias{R6_hrp2_mod}
\title{R6 Class for hrp2 selection model}
\description{
R6 Class for hrp2 selection model

R6 Class for hrp2 selection model
}
\details{
Model for predicting the selective coefficient for hrp2 deletions
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-hrp2_mod-new}{\code{R6_hrp2_mod$new()}}
\item \href{#method-hrp2_mod-predict}{\code{R6_hrp2_mod$predict()}}
\item \href{#method-hrp2_mod-predict_s}{\code{R6_hrp2_mod$predict_s()}}
\item \href{#method-hrp2_mod-predict_err}{\code{R6_hrp2_mod$predict_err()}}
\item \href{#method-hrp2_mod-predict_t}{\code{R6_hrp2_mod$predict_t()}}
\item \href{#method-hrp2_mod-predict_f2}{\code{R6_hrp2_mod$predict_f2()}}
\item \href{#method-hrp2_mod-add_model}{\code{R6_hrp2_mod$add_model()}}
\item \href{#method-hrp2_mod-add_model_weight}{\code{R6_hrp2_mod$add_model_weight()}}
\item \href{#method-hrp2_mod-add_err_model}{\code{R6_hrp2_mod$add_err_model()}}
\item \href{#method-hrp2_mod-add_err_model_weight}{\code{R6_hrp2_mod$add_err_model_weight()}}
\item \href{#method-hrp2_mod-get_models}{\code{R6_hrp2_mod$get_models()}}
\item \href{#method-hrp2_mod-get_error_models}{\code{R6_hrp2_mod$get_error_models()}}
\item \href{#method-hrp2_mod-get_model_weights}{\code{R6_hrp2_mod$get_model_weights()}}
\item \href{#method-hrp2_mod-get_error_model_weights}{\code{R6_hrp2_mod$get_error_model_weights()}}
\item \href{#method-hrp2_mod-get_data}{\code{R6_hrp2_mod$get_data()}}
\item \href{#method-hrp2_mod-set_models}{\code{R6_hrp2_mod$set_models()}}
\item \href{#method-hrp2_mod-set_error_models}{\code{R6_hrp2_mod$set_error_models()}}
\item \href{#method-hrp2_mod-set_model_weights}{\code{R6_hrp2_mod$set_model_weights()}}
\item \href{#method-hrp2_mod-set_error_model_weights}{\code{R6_hrp2_mod$set_error_model_weights()}}
\item \href{#method-hrp2_mod-set_data}{\code{R6_hrp2_mod$set_data()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hrp2_mod-new"></a>}}
\if{latex}{\out{\hypertarget{method-hrp2_mod-new}{}}}
\subsection{Method \code{new()}}{
Create a new hrp2 selection model using emulators.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_hrp2_mod$new(
  models = list(),
  err_models = list(),
  model_weights = list(),
  err_model_weights = list(),
  data = data.frame()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{models}}{Model for prediction seletion coefficient}

\item{\code{err_models}}{Model for predicting sd error for predictions}

\item{\code{data}}{Data used originally for model training}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{hrp2_mod} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hrp2_mod-predict"></a>}}
\if{latex}{\out{\hypertarget{method-hrp2_mod-predict}{}}}
\subsection{Method \code{predict()}}{
Predict median and 95\% selection coefficients and threshold times
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_hrp2_mod$predict(
  Micro.2.10,
  ft,
  microscopy.use,
  rdt.nonadherence,
  fitness,
  rdt.det
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Micro.2.10}}{Microscopy prevalence in 2-10 year olds}

\item{\code{ft}}{Treatment seeking rate}

\item{\code{microscopy.use}}{Proportion of total testing by microscopy}

\item{\code{rdt.nonadherence}}{Probability of treating negative indiviudal}

\item{\code{fitness}}{Relative fitness of hrp2 deleted parasite}

\item{\code{rdt.det}}{Chance of hrp2 deleted parasite yielding positive test}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data.frame of selection coefficients and times from 1\% to 5\%
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hrp2_mod-predict_s"></a>}}
\if{latex}{\out{\hypertarget{method-hrp2_mod-predict_s}{}}}
\subsection{Method \code{predict_s()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_hrp2_mod$predict_s(dat)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hrp2_mod-predict_err"></a>}}
\if{latex}{\out{\hypertarget{method-hrp2_mod-predict_err}{}}}
\subsection{Method \code{predict_err()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_hrp2_mod$predict_err(dat)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hrp2_mod-predict_t"></a>}}
\if{latex}{\out{\hypertarget{method-hrp2_mod-predict_t}{}}}
\subsection{Method \code{predict_t()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_hrp2_mod$predict_t(dat, f1, f2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hrp2_mod-predict_f2"></a>}}
\if{latex}{\out{\hypertarget{method-hrp2_mod-predict_f2}{}}}
\subsection{Method \code{predict_f2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_hrp2_mod$predict_f2(dat, f1, t)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hrp2_mod-add_model"></a>}}
\if{latex}{\out{\hypertarget{method-hrp2_mod-add_model}{}}}
\subsection{Method \code{add_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_hrp2_mod$add_model(model, model_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hrp2_mod-add_model_weight"></a>}}
\if{latex}{\out{\hypertarget{method-hrp2_mod-add_model_weight}{}}}
\subsection{Method \code{add_model_weight()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_hrp2_mod$add_model_weight(weight, model_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hrp2_mod-add_err_model"></a>}}
\if{latex}{\out{\hypertarget{method-hrp2_mod-add_err_model}{}}}
\subsection{Method \code{add_err_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_hrp2_mod$add_err_model(err_model, model_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hrp2_mod-add_err_model_weight"></a>}}
\if{latex}{\out{\hypertarget{method-hrp2_mod-add_err_model_weight}{}}}
\subsection{Method \code{add_err_model_weight()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_hrp2_mod$add_err_model_weight(weight, model_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hrp2_mod-get_models"></a>}}
\if{latex}{\out{\hypertarget{method-hrp2_mod-get_models}{}}}
\subsection{Method \code{get_models()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_hrp2_mod$get_models()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hrp2_mod-get_error_models"></a>}}
\if{latex}{\out{\hypertarget{method-hrp2_mod-get_error_models}{}}}
\subsection{Method \code{get_error_models()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_hrp2_mod$get_error_models()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hrp2_mod-get_model_weights"></a>}}
\if{latex}{\out{\hypertarget{method-hrp2_mod-get_model_weights}{}}}
\subsection{Method \code{get_model_weights()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_hrp2_mod$get_model_weights()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hrp2_mod-get_error_model_weights"></a>}}
\if{latex}{\out{\hypertarget{method-hrp2_mod-get_error_model_weights}{}}}
\subsection{Method \code{get_error_model_weights()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_hrp2_mod$get_error_model_weights()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hrp2_mod-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-hrp2_mod-get_data}{}}}
\subsection{Method \code{get_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_hrp2_mod$get_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hrp2_mod-set_models"></a>}}
\if{latex}{\out{\hypertarget{method-hrp2_mod-set_models}{}}}
\subsection{Method \code{set_models()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_hrp2_mod$set_models(models)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hrp2_mod-set_error_models"></a>}}
\if{latex}{\out{\hypertarget{method-hrp2_mod-set_error_models}{}}}
\subsection{Method \code{set_error_models()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_hrp2_mod$set_error_models(err_models)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hrp2_mod-set_model_weights"></a>}}
\if{latex}{\out{\hypertarget{method-hrp2_mod-set_model_weights}{}}}
\subsection{Method \code{set_model_weights()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_hrp2_mod$set_model_weights(model_weights)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hrp2_mod-set_error_model_weights"></a>}}
\if{latex}{\out{\hypertarget{method-hrp2_mod-set_error_model_weights}{}}}
\subsection{Method \code{set_error_model_weights()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_hrp2_mod$set_error_model_weights(err_model_weights)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hrp2_mod-set_data"></a>}}
\if{latex}{\out{\hypertarget{method-hrp2_mod-set_data}{}}}
\subsection{Method \code{set_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_hrp2_mod$set_data(data)}\if{html}{\out{</div>}}
}

}
}
