% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_figs}
\alias{save_figs}
\title{Save Figures}
\usage{
save_figs(
  name,
  fig,
  width = 6,
  height = 6,
  plot_dir = file.path(here::here(), "analysis/plots")
)
}
\arguments{
\item{name}{Name of figure}

\item{fig}{ggplot or similar figure object}

\item{width}{Width of plot in inches. Default = 6}

\item{height}{Height of plot in inches. Default = 6}

\item{plot_dir}{Plotting directory. Defaults to "analysis/plots"}
}
\description{
Save Figures
}
